requirements.txt
pandas
sqlalchemy
psycopg2-binary
requests
beautifulsoup4
lxml
